+++ app/controllers/application_controller.rb
@@ -33,12 +33,11 @@ class ApplicationController < ActionController::Base
 
   private
 
-    def empty_response_or_root_path(path = root_path)
-      respond_to do |format|
-        format.html { redirect_to path }
-        format.js { render nothing: true }
-        format.json { head :no_content }
-      end
+  def empty_response_or_root_path(path = root_path)
+    respond_to do |format|
+      format.html { redirect_to path }
+      format.js { render nothing: true }
+      format.json { head :no_content }
     end
-
+  end
 end