+++ app/models/template.rb
@@ -4,6 +4,9 @@ class Template < ActiveRecord::Base
   # Concerns
   include Deletable
 
+  attr_accessor :item_hashes
+  after_save :set_items
+
   # Named Scopes
 
   # Model Validations
@@ -12,6 +15,19 @@ class Template < ActiveRecord::Base
 
   # Model Relationships
   belongs_to :user
+  has_many :template_items, -> { order :position }
 
   # Template Methods
+
+  private
+
+  def set_items
+    return unless item_hashes && item_hashes.is_a?(Array)
+    template_items.destroy_all
+    item_hashes.each_with_index do |hash, index|
+      charge_type = user.charge_types.find_by_id hash[:charge_type_id]
+      decimal_amount = hash[:decimal_amount].to_s.gsub(/[\s$,]/, '')
+      template_items.create(charge_type_id: charge_type.id, position: index, name: hash[:name], decimal_amount: decimal_amount, description: hash[:description]) if charge_type
+    end
+  end
 end